PigSpy_Portable/requirements.txt
pyaudio>=0.2.11
speechrecognition>=3.8.1
requests>=2.28.1
aiohttp>=3.8.3
plyer>=2.1.0
winotify>=1.1.0
textual>=0.18.0
numpy>=1.24.1
sounddevice>=0.4.5
webrtcvad>=2.0.10
pydub>=0.25.1
python-vlc>=3.0.18
